requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]>=4.0.0
certifi
scikit-learn
mlflow==1.30.0
dagshub
pyaml
# -e .